tests and holiday config

+++ generate-holiday-config/index.js
@@ -1,36 +1,59 @@
 const fs = require('fs');
 const Holidays = require('date-holidays');
 
-const hd = new Holidays();
+//  change these values to generate a different country's holidays, or more years
+const COUNTRY_CODE = 'US';
+const YEARS_TO_GENERATE = 5;
+const FILENAME = 'holiday-config.json';
 
-const countryCode = 'US';
-const today = new Date();
-const currentYear = today.getFullYear();
-const yearsToGenerate = 5;
-const filename = 'holiday-config.json';
+// generates a holiday object for the given country and years
+function generateHolidayConfig(countryCode = COUNTRY_CODE, yearsToGenerate = YEARS_TO_GENERATE, startYear = new Date().getFullYear()) {
+  const hd = new Holidays();
+  hd.init(countryCode);
+  const holidays = {};
 
+  for (let i = 0; i < yearsToGenerate; i++) {
+    const year = startYear + i;
+    const yearHolidays = hd.getHolidays(year);
+    yearHolidays.forEach(({ name, date }) => {
+      const [year, month, day] = date.split(' ')[0].split('-');
+      const formattedHoliday = {
+        title: name,
+        date: {
+          year, month, day,
+        }
+      }
 
-hd.init(countryCode);
-const holidays = [];
-
-for (let i = 0; i < yearsToGenerate; i++) {
-  const year = currentYear + i;
-  const yearHolidays = hd.getHolidays(year);
-  yearHolidays.forEach((holiday) => {
-    const dateString = holiday.date.split(' ')[0];
-    const month = dateString.split('-')[1];
-    const day = dateString.split('-')[2];
-    holidays.push({
-      title: holiday.name,
-      date: {
-        year,
-        month,
-        day,
-      },
+      if (!holidays.hasOwnProperty(year)) {
+        holidays[year] = [formattedHoliday];
+      } else if (!Array.isArray(holidays[year])) {
+        console.log(formattedHoliday)
+        throw new Error(`Error: Invalid holiday-config.json: holidays[${year}] is not an array`);
+      } else {
+        holidays[year].push(formattedHoliday);
+      }
     });
-  });
+  }
+  return holidays
 }
-fs.unlinkSync(filename); // delete file if it exists
-fs.writeFileSync(filename, JSON.stringify(holidays, null, 2));
 
-console.log(`Configuration file saved to ${filename}`);
+// writes the holiday object to a file
+function saveHolidayConfigToDisk(holidays, filename = FILENAME) {
+  try {
+    if (fs.existsSync(filename)) {
+      fs.unlinkSync(filename);
+    }
+    fs.writeFileSync(filename, JSON.stringify(holidays, null, 2));
+  } catch (err) {
+    console.error(`Error writing to ${filename}: ${err}`)
+  }
+
+  console.log(`Configuration file saved to ${filename}`);
+}
+
+const holidays = generateHolidayConfig();
+saveHolidayConfigToDisk(holidays);
+
+module.exports = {
+  generateHolidayConfig, saveHolidayConfigToDisk
+};
